Decouple the register fn from the useForm hook
